mark in the Product class, the fild with anotation ManyToOne as JsonIgnore

+++ src/main/java/org/turkovaleksey/eshop/repository/api/PhotoRepository.java
@@ -2,10 +2,8 @@ package org.turkovaleksey.eshop.repository.api;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.turkovaleksey.eshop.repository.model.photo.Photo;
-
-import java.util.List;
+import org.turkovaleksey.eshop.repository.model.product.Photo;
 
 @Repository
-public interface PhotoRepository extends JpaRepository<Photo, Integer> {
+public interface PhotoRepository extends JpaRepository<Photo, Long> {
 }
